Refactor native
It should now be much more easy to make changes, since less code repetition.
However, understandability is worse now
