+++ darts/src/darts/__init__.py
@@ -2,7 +2,4 @@
 
 from importlib.metadata import version
 
-from darts.native import run_native_planet_pipeline as run_native_planet_pipeline
-from darts.native import run_native_sentinel2_pipeline as run_native_sentinel2_pipeline
-
 __version__ = version("darts-nextgen")
